refactor(agents): split subagent spawn helpers
